move dialogs to separated file
